+++ NEWS.md
@@ -6,8 +6,27 @@ Note that as of Julia v1.5, in order to see deprecation warnings you will need t
 
 After we release v1 of the package, we will start following [semantic versioning](https://semver.org).
 
+ITensors v0.3.2 Release Notes
+=============================
+
+Bugs:
+
+Enhancements:
+
+- Introduce `set_nsite!` generic `AbstractProjMPO` function (#894)
+- Factorize out `contract(::ProjMPO, ::ITensor)` (#893)
+
+ITensors v0.3.1 Release Notes
+=============================
+
+Bugs:
+
+Enhancements:
+
+- Introduce `Algorithm` type for selecting algorithm backends (#886)
+
 ITensors v0.3.0 Release Notes
-==============================
+=============================
 
 Bugs: